requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
alembic
httpx
openai
langchain
langchain-openai
pyyaml
psycopg2-binary
PyJWT[crypto]
cryptography
ecdsa
